eLearning/requirements.txt
aioredis==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
billiard==4.2.0
bootstrap4==0.1.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
channels==3.0.4
channels-redis==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==42.0.5
daphne==3.0.2
Django==5.0.2
django-bootstrap4==24.1
djangorestframework==3.14.0
hiredis==2.3.2
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
kombu==5.3.5
msgpack==1.0.8
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg==3.1.18
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.0.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
redis==5.0.2
service-identity==24.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.2
